Fix error after zipkin internal renaming
